feat: 플레이어는 딜러 이름을 사용할 수 없도록 예외처리
