+++ src/main/java/domain/Player.java
@@ -1,63 +1,23 @@
 package domain;
 
-import java.util.List;
-import java.util.Objects;
-
-public class Player {
-    private final Name name;
-    private final Hand hand;
-
+public class Player extends Participant {
     public Player(final Name name, final Hand hand) {
-        this.name = name;
-        this.hand = hand;
+        super(name, hand);
+        validateName(name);
     }
 
     public Player(final Name name) {
         this(name, new Hand());
     }
 
-    public void dealCards(final List<Card> cards) {
-        hand.addAll(cards);
+    private void validateName(final Name name) {
+        if (name.getName().equals("딜러")) {
+            throw new IllegalArgumentException();
+        }
     }
 
-    public void dealCard(final Card card) {
-        hand.add(card);
-    }
-
-    public int score() {
-        return hand.score().toInt();
-    }
-
-    public boolean isBust() {
-        return hand.isBust();
-    }
-
-    public boolean isNotBust() {
+    @Override
+    public boolean canHit() {
         return hand.isNotBust();
-    }
-
-    public String name() {
-        return name.getName();
-    }
-
-    public List<Card> hand() {
-        return hand.cards();
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        final Player player = (Player) o;
-        return Objects.equals(name, player.name);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(name);
     }
 }
